# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aura Health API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cycle/entries:
    get:
      operationId: getCycleEntries
      tags: [cycle]
      summary: Get all cycle log entries
      responses:
        "200":
          description: List of cycle entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CycleEntry"
    post:
      operationId: createCycleEntry
      tags: [cycle]
      summary: Log a cycle entry (period start, ovulation, etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CycleEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CycleEntry"
  /cycle/entries/{id}:
    patch:
      operationId: updateCycleEntry
      tags: [cycle]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Update a cycle entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CycleEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CycleEntry"
    delete:
      operationId: deleteCycleEntry
      tags: [cycle]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Delete a cycle entry
      responses:
        "204":
          description: Deleted
...
  /marketplace/products/{id}:
    get:
      operationId: getProduct
      tags: [marketplace]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Get a single product
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /marketplace/cart:
    get:
      operationId: getCart
      tags: [marketplace]
      summary: Get cart items
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItem"
    post:
      operationId: addToCart
      tags: [marketplace]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
  /marketplace/cart/{id}:
    delete:
      operationId: removeFromCart
      tags: [marketplace]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Remove item from cart
      responses:
        "204":
          description: Removed
...
  /zenzone/scores:
    get:
      operationId: getZenScores
      tags: [zenzone]
      summary: Get user's Zen Zone game scores
      responses:
        "200":
          description: Scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ZenScore"
    post:
      operationId: saveZenScore
      tags: [zenzone]
      summary: Save a Zen Zone game score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZenScoreInput"
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZenScore"
...
  /chat/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get chat history
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the Aura myth-buster chatbot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /questionnaire/responses:
    get:
      operationId: getQuestionnaireResponses
      tags: [questionnaire]
      summary: Get saved questionnaire responses
      responses:
        "200":
          description: Responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestionnaireResponse"
    post:
      operationId: saveQuestionnaireResponse
      tags: [questionnaire]
      summary: Save a questionnaire swipe response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionnaireResponseInput"
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionnaireResponse"
  /partner/sync:
    get:
      operationId: getPartnerSync
      tags: [partner]
      summary: Get partner sync settings and shared data
      responses:
        "200":
          description: Partner sync info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerSync"
    post:
      operationId: updatePartnerSync
      tags: [partner]
      summary: Update partner sync preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerSyncInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerSync"
...
components:
  schemas:
...
    CycleEntry:
      type: object
      required: [id, date, type]
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        type:
          type: string
          enum: [period_start, period_end, ovulation, fertile_window, spotting]
        flow:
          type: ["string", "null"]
          enum: [light, medium, heavy, null]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CycleEntryInput:
      type: object
      required: [date, type]
      properties:
        date:
          type: string
          format: date
        type:
          type: string
          enum: [period_start, period_end, ovulation, fertile_window, spotting]
        flow:
          type: ["string", "null"]
          enum: [light, medium, heavy, null]
        notes:
          type: ["string", "null"]
    CycleEntryUpdate:
      type: object
      properties:
        date:
          type: string
          format: date
        type:
          type: string
          enum: [period_start, period_end, ovulation, fertile_window, spotting]
        flow:
          type: ["string", "null"]
          enum: [light, medium, heavy, null]
        notes:
          type: ["string", "null"]
...
    Product:
      type: object
      required: [id, name, description, price, category, imageUrl, ecoRating, inStock]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
          enum: [pads, tampons, cups, heat_patches, chocolates, supplements]
        imageUrl:
          type: string
        ecoRating:
          type: integer
        inStock:
          type: boolean
        badge:
          type: ["string", "null"]
    CartItem:
      type: object
      required: [id, productId, quantity, product]
      properties:
        id:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
...
    ZenScore:
      type: object
      required: [id, score, level, timeSeconds, createdAt]
      properties:
        id:
          type: integer
        score:
          type: integer
        level:
          type: integer
        timeSeconds:
          type: integer
        pointsEarned:
          type: integer
        createdAt:
          type: string
          format: date-time
    ZenScoreInput:
      type: object
      required: [score, level, timeSeconds]
      properties:
        score:
          type: integer
        level:
          type: integer
        timeSeconds:
          type: integer
...
    ChatMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        language:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        language:
          type: ["string", "null"]
    QuestionnaireResponse:
      type: object
      required: [id, questionId, answer, createdAt]
      properties:
        id:
          type: integer
        questionId:
          type: string
        answer:
          type: string
        createdAt:
          type: string
          format: date-time
    QuestionnaireResponseInput:
      type: object
      required: [questionId, answer]
      properties:
        questionId:
          type: string
        answer:
          type: string
    PartnerSync:
      type: object
      required: [id, enabled, sharePhase, shareSymptoms, shareMood, partnerEmail]
      properties:
        id:
          type: integer
        enabled:
          type: boolean
        sharePhase:
          type: boolean
        shareSymptoms:
          type: boolean
        shareMood:
          type: boolean
        partnerEmail:
          type: ["string", "null"]
        lastSynced:
          type: ["string", "null"]
          format: date-time
    PartnerSyncInput:
      type: object
      properties:
        enabled:
          type: boolean
        sharePhase:
          type: boolean
        shareSymptoms:
          type: boolean
        shareMood:
          type: boolean
        partnerEmail:
          type: ["string", "null"]
...